Log warning when wireless plugin is used on unsupported platform
